Conflicts:
	apic_ml2/neutron/plugins/ml2/drivers/cisco/apic/config.py
	apic_ml2/neutron/plugins/ml2/drivers/cisco/apic/mechanism_apic.py
	apic_ml2/neutron/services/l3_router/l3_apic.py
	apic_ml2/neutron/tests/unit/ml2/drivers/cisco/apic/test_cisco_apic_mechanism_driver.py
	apic_ml2/neutron/tests/unit/services/l3_router/test_l3_apic_plugin.py
	debian/build-deb.sh
	debian/changelog.in
	requirements.txt
	rpm/build-rpm.sh
	rpm/neutron-ml2-driver-apic.spec.in
	setup.cfg

+++ debian/build-deb.sh
@@ -2,17 +2,26 @@
 # Should be run from the root of the source tree
 # Set env var REVISION to overwrite the 'revision' field in version string
 
+if [ ! -d debian ]; then
+   echo "Directory 'debian' not found"
+   exit 1
+fi
+if [ ! -f debian/changelog.in ]; then
+   echo "Debian changelog file not found"
+   exit 1
+fi
 BUILD_DIR=${BUILD_DIR:-`pwd`/debbuild}
 mkdir -p $BUILD_DIR
 rm -rf $BUILD_DIR/*
 NAME=`python setup.py --name`
-VERSION=`python setup.py --version`
+VERSION_PY=`python setup.py --version`
+VERSION=${VERSION_PY/.dev/~dev}
 REVISION=${REVISION:-1}
 python setup.py sdist --dist-dir $BUILD_DIR
-SOURCE_FILE=${NAME}-${VERSION}.tar.gz
+SOURCE_FILE=${NAME}-${VERSION_PY}.tar.gz
 tar -C $BUILD_DIR -xf $BUILD_DIR/$SOURCE_FILE
-SOURCE_DIR=$BUILD_DIR/${NAME}-${VERSION}
-
+SOURCE_DIR=$BUILD_DIR/${NAME}-${VERSION_PY}
+cp -H -r debian $SOURCE_DIR/
 sed -e "s/@VERSION@/$VERSION/" -e "s/@REVISION@/$REVISION/" ${SOURCE_DIR}/debian/changelog.in > ${SOURCE_DIR}/debian/changelog
 
 mv $BUILD_DIR/$SOURCE_FILE $BUILD_DIR/${NAME}_${VERSION}.orig.tar.gz
